+++ CHANGES.md
@@ -3,6 +3,7 @@ Changes
 
 0.0.7-dev
 * New Extra/Cursor: Set mouse cursor contextually. Can auto-hide cursor with time out.
+* Init `three.Time.now` on install.
 
 0.0.6
 * Core: 'ready' event now always fires on hot install